requirements.txt
beautifulsoup4>=4.12.0
aiohttp>=3.10.0
python-dotenv>=1.0.0
pytest>=8.0.0

